Modified how pin and value variables are intialized and also the reading process
The pin variable are set in two different arrays one for digital and one for the analog pins. The value read from the pins are stored in a struct which has two arrays one being int for storing analog values and the other being a bool for strong the digital values. The reason for doing this is because it helps reduce the repetation of declaring, intializing and reading all the values from different pins